+++ src/apps/music/mod.rs
@@ -1,3 +1,0 @@
-mod spotatui;
-
-pub use spotatui::Spotatui;